Cmma f32 accumulate + cubek-tile module reorg
Tile-DSL Cmma routine now accumulates f16/bf16 outputs in f32 (cast down
to the output on drain), closing the ~0.008% precision drift vs the legacy
SimpleCyclicCmma it ports.

- Tile::promote is generic over a second register type EA; the accumulator
  fragment mirrors it and is zero-filled (beta=0), replacing the old
  same-type gmem-load init. Cross-type epilogue Tile::drain_cast_into
  (cmma::cast::<EA,Out> + store); copy_from stays same-type.
- The routine mandates the acc type (from_globals upgrade) rather than
  trusting dtypes.acc_register, which is f16 on the bench/correctness
  paths (from_single_dtype does not upgrade). Keys both the MmaConfig
  lookup (cd_type) and the kernel's EA arg.
- Selector stage-depth cap is now a byte budget (stage_k * operand_bytes),
  which unifies the tuning: f16 -> sk32, f32 -> sk16. f32 at the old sk64
  sat past the smem residency cliff.

Metal, square_2x4096, vs legacy: f16 ~103%, f32 ~140%; precision matches.
Also folds in the concurrent cubek-tile module reorg (ops/, physical/,
space/partition/, staging/, tile/view/).
