+++ crates/cubek-tile/src/lib.rs
@@ -2,31 +2,17 @@
 #![allow(dead_code)]
 
 mod fold;
-mod layout;
-mod load;
-mod matmul;
-mod partitioner;
-mod quantization;
-mod resident;
-mod scalar;
-mod softmax;
+mod ops;
+mod physical;
 mod space;
-mod stage;
+mod staging;
 mod tile;
-mod vec_tensor;
-mod view;
 
 // `Axis`/`MAX_AXES` and `ConcreteLayout` are the storage-layout vocabulary; clients reach them
 // through `cubek_tile::{Axis, ...}`.
 pub use fold::*;
-pub use layout::*;
-pub use load::*;
-pub use partitioner::*;
-pub use quantization::*;
-pub use scalar::*;
-pub use softmax::*;
+pub use ops::*;
+pub use physical::*;
 pub use space::*;
-pub use stage::*;
+pub use staging::*;
 pub use tile::*;
-pub use vec_tensor::*;
-pub use view::*;
