+++ crates/cubek-tile/src/ops/matmul/instruction/base.rs
@@ -5,7 +5,8 @@
 
 use cubecl::prelude::*;
 
-use crate::{matmul::instruction::register::mma_register_memory, *};
+use crate::*;
+use super::register::mma_register_memory;
 
 /// The leaf contraction `acc += lhs · rhs`. Dispatch is dynamic on the accumulator's comptime
 /// storage config
