+++ crates/cubek-tile/src/space/mod.rs
@@ -1,7 +1,11 @@
+//! The logical coordinate space a tile lives in, and how a level of it splits ([`partition`]).
+
 mod axis;
-mod base;
+mod partition;
 mod region;
+mod space;
 
 pub use axis::*;
-pub use base::*;
+pub use partition::*;
 pub use region::*;
+pub use space::*;
