+++ crates/cubek-tile/src/space/partition/geometry.rs
