+++ crates/cubek-tile/src/space/partition/mod.rs
@@ -1,14 +1,14 @@
 //! How a level of the space splits, and the [`Walk`] it produces.
 
-mod base;
 mod distribution;
-mod launch_geometry;
+mod geometry;
 mod launcher;
+mod partitioner;
 mod tiling;
 mod walk;
 mod walk_order;
 
-pub use base::*;
+pub use partitioner::*;
 pub use distribution::*;
 pub use launcher::*;
 pub use tiling::*;
