+++ crates/cubek-tile/src/tile/view/masked.rs
